ランディングページをコンポーネントに分離
